requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
asyncpg
pydantic
python-dotenv